moved some things around and refactored the earlier parts of the piecework literature review
i didn't bother to change the narrative of the sections yet, in part because i don't really know how to rewrite it from the chronological ordering, especially going from the industrial work->farmwork sections, since taylorism was sort of a directional bridge. maybe i could write it so it's more like 'where did taylorism come from though? for that, we turn to farmwork...'?